Module pyrev_app::prelude

Re-exports

pub use crate::app::App;

pub use crate::plugin::Plugin;

pub use crate::plugin::*;

Macros

arg

Create an Arg from a usage string.

command

Allows you to build the Command instance from your Cargo.toml at compile time.

error

info

value_parser

Select a [ValueParser] implementation from the intended type

warn

Structs

Arg

The abstract representation of a command line argument. Used to set all the options and relationships that define a valid argument for the program.

ArgMatches

Container for parse results.

Cli

Command

Build a command-line interface.

DecompiledCode

Local

OpcodeInstruction

OrderMap

一个简单的有序字典

ResMut

TraceBack

Enums

ArgAction

Behavior of arguments when they are encountered while parsing

Opcode

Traits

Colorize

The trait that enables something to be given color.

Decompiler

ExprParser

IStream

OStream

OpcodeParser

Functions

get_trace

Type Aliases

CodeObject

CodeObjectMap

LineNumber

ObjectMark

Result